requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
asyncpg==0.29.0
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
click==8.1.7
databases==0.9.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
isort==5.13.2
magic-filter==1.0.12
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
yarl==1.13.1
